# modules/docs/swagger/enrollment-api.yaml
openapi: 3.0.3
info:
  title: Facial Enrollment System API
  description: |
    API specification for the Facial Enrollment System, covering:
    - IoT Broker Data Service (device management, enrollment publishing)
    - Facial Enrollment Modal (embedded in Bridge server)
    - GPU Vectorizer API (embedding generation)
    
    ## System Architecture
    
    ```
    ┌─────────────────────────────────────────────────────────────────────────┐
    │                          BRIDGE SERVER                                   │
    │  ┌───────────────────────────────────────────────────────────────────┐  │
    │  │                    Facial Enrollment Modal                         │  │
    │  │  (React Frontend + Node.js API Server)                            │  │
    │  │                                                                    │  │
    │  │  Receives:                                                         │  │
    │  │    ← POST from Employee Management Web Page (opens modal)          │  │
    │  │                                                                    │  │
    │  │  Sends:                                                            │  │
    │  │    → IoT Broker: Publish enrollment to edge devices                │  │
    │  │    → GPU Server: Generate embeddings & VLM analysis                │  │
    │  └───────────────────────────────────────────────────────────────────┘  │
    └─────────────────────────────────────────────────────────────────────────┘
...
    
...
servers:
...
paths:
...
  /api/vectorizer/generate:
    post:
      tags:
        - Vectorizer
      summary: Generate facial embedding
      description: |
        Process captured images and generate facial embedding on the AWS GPU server.
        
        Returns:
        - `enrollmentProcessedFile`: Base64 Float32Array (512 dimensions)
        - `enrollmentPictureThumbnail`: Base64 JPEG 128x128
      operationId: generateEmbedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VectorizerRequest'
      responses:
        '200':
          description: Embedding generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VectorizerResponse'
        '400':
          description: Invalid request - missing images or invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unable to process images - no face detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
...
  schemas:
...
    PoseName:
      type: string
      description: Capture pose name
      enum:
        - front
        - left
        - right
        - up
        - down
      example: front

    # ---- Device Schemas ----
...
    VectorizerImage:
      type: object
      required:
        - pose
        - data
      properties:
        pose:
          $ref: '#/components/schemas/PoseName'
        data:
          type: string
          description: Base64 encoded image data
          example: /9j/4AAQSkZJRg...
    VectorizerRequest:
      type: object
      required:
        - employee_id
        - images
        - function
      properties:
        employee_id:
          type: string
          example: EMP123456
        images:
          type: array
          minItems: 5
          maxItems: 5
          items:
            $ref: '#/components/schemas/VectorizerImage'
        function:
          type: string
          description: Processing function type
          enum:
            - facial_embedding
            - emotion_analysis
          example: facial_embedding
    VectorizerResponse:
      type: object
      required:
        - employee_id
        - enrollmentProcessedFile
        - embedding_dim
        - model
        - enrollmentPictureThumbnail
        - image_count
      properties:
        employee_id:
          type: string
          example: EMP123456
        enrollmentProcessedFile:
          type: string
          description: Base64 encoded Float32Array (512 dimensions for ArcFace)
          example: AACAPwAAgD8AAIA/...
        embedding_dim:
          type: integer
          description: Embedding vector dimension
          example: 512
        model:
          type: string
          description: Model used for embedding
          example: ArcFace
        enrollmentPictureThumbnail:
          type: string
          description: Base64 encoded JPEG 128x128 thumbnail
          example: /9j/4AAQSkZJRg...
        image_count:
          type: integer
          description: Number of images successfully processed
          example: 5

    # ---- Error Schema ----
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
          example: Employee enrollment not found
        code:
          type: string
          description: Error code for programmatic handling
          example: ENROLLMENT_NOT_FOUND
        details:
          type: object
          additionalProperties: true
          description: Additional error context

  # ===========================================
  # Security Schemes
  # ===========================================
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from WFM authentication

    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant identifier for service-to-service calls
security:
  - BearerAuth: []